+++ src/main/java/be/ugent/knows/idlabFunctions/state/SimpleInMemoryMapState.java
@@ -5,6 +5,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.*;
 import java.util.*;
+import java.util.stream.Collectors;
 
 /**
  * <p>Copyright 2022 IDLab (Ghent University - imec)</p>
@@ -92,6 +93,70 @@ public class SimpleInMemoryMapState implements MapState {
         }
     }
 
+    @Override
+    public void replace(String stateFilePath, String key, List<String> value) {
+        Map<String, List<String>> map = stateFileToMap.computeIfAbsent(stateFilePath, mapKey -> {
+            // first check if file exists and try to load map
+            File stateFile = new File(stateFilePath);
+            Map<String, List<String>> newMap = new HashMap<>();
+            if (stateFile.exists() && stateFile.isFile() && stateFile.canRead()) {
+                try (ObjectInputStream in = new ObjectInputStream(new BufferedInputStream(new FileInputStream(stateFilePath)))){
+                    newMap = (Map<String, List<String>>)in.readObject();
+                } catch (IOException | ClassNotFoundException e) {
+                    log.warn("Cannot load state map from file {}. Creating empty map!", stateFilePath);
+                }
+            }
+            return newMap;
+        });
+        List<String> values = map.computeIfAbsent(key, k -> new ArrayList<>());
+        values = value;
+    }
+
+    @Override
+    public boolean hasKey(String stateFilePath, String key) {
+        Map<String, List<String>> map = stateFileToMap.computeIfAbsent(stateFilePath, mapKey -> {
+            // first check if file exists and try to load map
+            File stateFile = new File(stateFilePath);
+            Map<String, List<String>> newMap = new HashMap<>();
+            if (stateFile.exists() && stateFile.isFile() && stateFile.canRead()) {
+                try (ObjectInputStream in = new ObjectInputStream(new BufferedInputStream(new FileInputStream(stateFilePath)))){
+                    newMap = (Map<String, List<String>>)in.readObject();
+                } catch (IOException | ClassNotFoundException e) {
+                    log.warn("Cannot load state map from file {}. Creating empty map!", stateFilePath);
+                }
+            }
+
+            return newMap;
+        });
+
+        List<String> values = map.computeIfAbsent(key, k -> new ArrayList<>(4));
+
+        if (values.isEmpty())
+            return false;
+
+        return true;
+    }
+
+    @Override
+    public Map<String, List<String>> getEntries(String stateFilePath) {
+        Map<String, List<String>> map = stateFileToMap.computeIfAbsent(stateFilePath, mapKey -> {
+            // first check if file exists and try to load map
+            File stateFile = new File(stateFilePath);
+            Map<String, List<String>> newMap = new HashMap<>();
+            if (stateFile.exists() && stateFile.isFile() && stateFile.canRead()) {
+                try (ObjectInputStream in = new ObjectInputStream(new BufferedInputStream(new FileInputStream(stateFilePath)))){
+                    newMap = (Map<String, List<String>>)in.readObject();
+                } catch (IOException | ClassNotFoundException e) {
+                    log.warn("Cannot load state map from file {}. Creating empty map!", stateFilePath);
+                }
+            }
+
+            return newMap;
+        });
+
+        return map;
+    }
+
     /**
      * Closes the state: state is persisted to disk before clearing the state in memory.
      */
@@ -135,5 +200,23 @@ public class SimpleInMemoryMapState implements MapState {
         } else {
             return 0;
         }
+    }
+
+    @Override
+    public void remove(String stateFilePath, String key) {
+        Map<String, List<String>> map = stateFileToMap.computeIfAbsent(stateFilePath, mapKey -> {
+            // first check if file exists and try to load map
+            File stateFile = new File(stateFilePath);
+            Map<String, List<String>> newMap = new HashMap<>();
+            if (stateFile.exists() && stateFile.isFile() && stateFile.canRead()) {
+                try (ObjectInputStream in = new ObjectInputStream(new BufferedInputStream(new FileInputStream(stateFilePath)))){
+                    newMap = (Map<String, List<String>>)in.readObject();
+                } catch (IOException | ClassNotFoundException e) {
+                    log.warn("Cannot load state map from file {}. Creating empty map!", stateFilePath);
+                }
+            }
+            return newMap;
+        });
+        map.remove(key);
     }
 }