+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## [0.3.2] - 2023-06-12
+
+### Fixed
+- IDLabFunctions: rename 'template' to 'iri'
+
+### Changed
+- IDLabFunctions: move state dir path resolving to separate method
+
+### Added
+- Map: add hasKey method
+- Map: add getEntries method
+- Map: add replace method
+- Map: add remove method
+- IDLabFunctions: add implicitCreate function
+- IDLabFunctions: add implicitUpdate function
+- IDLabFunctions: add implicitDelete function
+
 ## [0.3.1] - 2023-05-04
 
 ### Fixed
@@ -72,6 +91,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Code extracted from [RMLMapper](https://github.com/RMLio/rmlmapper-java)
 
+[0.3.2]: https://github.com/FnOio/idlab-functions-java/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/FnOio/idlab-functions-java/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/FnOio/idlab-functions-java/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/FnOio/idlab-functions-java/compare/v0.1.5...v0.2.0